+++ include/sacn/merge_receiver.h
@@ -145,7 +145,7 @@ typedef struct SacnMergeReceiverConfig
   /** The maximum number of sources this universe will listen to.  May be #SACN_RECEIVER_INFINITE_SOURCES.
       This parameter is ignored when configured to use static memory -- #SACN_RECEIVER_MAX_SOURCES_PER_UNIVERSE is used
      instead.*/
-  size_t source_count_max;
+  int source_count_max;
 } SacnMergeReceiverConfig;
 
 /** A default-value initializer for an SacnMergeReceiverConfig struct. */
